chore(): update api places request
