+++ src/controllers/places/index.js
@@ -1,0 +1,2 @@
+export * from "./place"
+export * from "./places"
