+++ src/pages/api/places/index.js
@@ -1,43 +1,4 @@
-import { PlaceModel } from "@/features/places/models"
-import { filterQueryPlace } from "@/features/places/schemas"
+import { PlacesController } from "@/controllers/places"
 import { handlerApi } from "@/utils/functions"
 
-const handler = handlerApi(async (req, res) => {
-	if (req.method === "GET") {
-		const { q, limit, page, ...query } = req.query
-		const places = q
-			? PlaceModel.find({
-					$text: {
-						$search: q,
-					},
-					...(await filterQueryPlace.parseAsync(query)),
-				})
-			: PlaceModel.find(await filterQueryPlace.parseAsync(query))
-
-		if (limit) {
-			return res.json(await places.limit(limit))
-		}
-
-		if (page) {
-			return res.json(await places.skip((limit - 1 || 9) * page))
-		}
-
-		return res.json(await places)
-	}
-
-	if (req.method === "POST") {
-		const { info, details } = req.body
-		const newPlace = new PlaceModel({
-			...info,
-			[info.building]: details,
-		})
-
-		await newPlace.save()
-
-		return res.json(newPlace)
-	}
-
-	return res.status(405).send("Method not allowed")
-})
-
-export default handler
+export default handlerApi(PlacesController)
