+++ src/utils/functions.js
@@ -1,4 +1,5 @@
 import { clsx } from "clsx"
+import Redis from "ioredis"
 import mongoose from "mongoose"
 import { twMerge } from "tailwind-merge"
 
@@ -12,24 +13,32 @@ export const cn = (...inputs) => twMerge(clsx(inputs))
 
 /**
  *
- * @param {Function | Promise<Function>} handle
+ * @param {Function | Promise<Function>} controller
  * @returns {Function}
  */
 
-export const handlerApi = handle => async (req, res) => {
+export const handlerApi = controller => async (req, res) => {
+	const redisClient = new Redis({
+		maxRetriesPerRequest: 3,
+	})
+
 	try {
 		await mongoose.connect(process.env.DATABASE_URL)
 
-		await handle(req, res)
+		if (!controller[req.method]) {
+			return res.status(405).send("Method not allowed")
+		}
+
+		return await controller[req.method](req, res, redisClient)
 	} catch (error) {
 		if (
 			error.message.includes("ECONNREFUSED") &&
 			error.message.includes("27017")
 		) {
-			res.status(500).json({ error: "Database connection failed" })
+			return res.status(500).json({ error: "Database connection failed" })
 		}
 
-		res.status(500).json({ error: "Unknown server error" })
+		return res.status(500).json({ error: "Unknown server error" })
 	}
 }
